+++ common/src/main/java/com/chefmooon/ubesdelight/common/registry/UbesDelightAdvancements.java
@@ -1,20 +1,23 @@
 package com.chefmooon.ubesdelight.common.registry;
 
 import com.chefmooon.ubesdelight.common.advancement.BakingMatTrigger;
+import com.chefmooon.ubesdelight.common.advancement.LeafFeastTrigger;
 import dev.architectury.injectables.annotations.ExpectPlatform;
+import net.minecraft.advancements.CriterionTrigger;
 
 import java.util.function.Supplier;
 
 public class UbesDelightAdvancements {
-//    public static final DeferredRegister<CriterionTrigger<?>> TRIGGERS = DeferredRegister.create(Registries.TRIGGER_TYPE, UbesDelight.MOD_ID);
-//    public static final Supplier<BakingMatTrigger> USE_BAKING_MAT = TRIGGERS.register("use_baking_mat", BakingMatTrigger::new);
-//
-//    public static void register() {
-//        TRIGGERS.register();
-//    }
+    public static final Supplier<BakingMatTrigger> USE_BAKING_MAT = registerTrigger("use_baking_mat", BakingMatTrigger::new);
+    public static final Supplier<LeafFeastTrigger> USE_LEAF_FEAST = registerTrigger("use_leaf_feast", LeafFeastTrigger::new);
 
     @ExpectPlatform
-    public static Supplier<BakingMatTrigger> getBakingMatTrigger() {
+    public static <T extends CriterionTrigger<?>> Supplier<T> registerTrigger(String name, Supplier<T> triggerSupplier) {
         throw new AssertionError();
     }
+
+    public static void init() {
+
+    }
+
 }