+++ fabric/src/main/java/com/chefmooon/ubesdelight/data/fabric/DataGenerationFabric.java
@@ -15,5 +15,6 @@ public class DataGenerationFabric implements DataGeneratorEntrypoint {
         pack.addProvider(ModelGenerator::new);
         pack.addProvider(RecipeGenerator::new);
         pack.addProvider(LootTableGenerator::new);
+        pack.addProvider(AdvancmentGenerator::new);
     }
 }