+++ neoforge/src/main/java/com/chefmooon/ubesdelight/common/registry/neoforge/UbesDelightAdvancementsImpl.java
@@ -2,6 +2,7 @@ package com.chefmooon.ubesdelight.common.registry.neoforge;
 
 import com.chefmooon.ubesdelight.UbesDelight;
 import com.chefmooon.ubesdelight.common.advancement.BakingMatTrigger;
+import com.chefmooon.ubesdelight.common.advancement.LeafFeastTrigger;
 import net.minecraft.advancements.CriterionTrigger;
 import net.minecraft.core.registries.Registries;
 import net.neoforged.bus.api.IEventBus;
@@ -11,13 +12,12 @@ import java.util.function.Supplier;
 
 public class UbesDelightAdvancementsImpl {
     public static final DeferredRegister<CriterionTrigger<?>> TRIGGERS = DeferredRegister.create(Registries.TRIGGER_TYPE, UbesDelight.MOD_ID);
-    public static final Supplier<BakingMatTrigger> USE_BAKING_MAT = TRIGGERS.register("use_baking_mat", BakingMatTrigger::new);
+
+    public static <T extends CriterionTrigger<?>> Supplier<T> registerTrigger(String name, Supplier<T> triggerSupplier) {
+        return TRIGGERS.register(name, triggerSupplier);
+    }
 
     public static void register(IEventBus eventBus) {
         TRIGGERS.register(eventBus);
-    }
-
-    public static Supplier<BakingMatTrigger> getBakingMatTrigger() {
-        return USE_BAKING_MAT;
     }
 }